requirements.txt
opencv-python==3.3.0.9
opencv-contrib-python==3.4.0.12
scikit-image==0.13.1
imutils==0.4.6
scipy==1.1.0
Pillow==4.2.1